convert to no_std crate
Change uses of std data structures to equivalents from core, alloc,
and hashbrown
