+++ Cargo.toml
@@ -7,5 +7,8 @@ description = "Reads and parses Linux device tree images"
 repository = "https://github.com/mbr/device_tree-rs"
 documentation = "https://mbr.github.io/device_tree-rs/device_tree/"
 
+[dependencies]
+hashbrown = "0.13"
+
 [features]
 string-dedup = [] # Requires std
