+++ src/lib.rs
@@ -18,10 +18,11 @@
 //!
 //! # Examples
 //!
-//! ```ignore
+//! ```rust
+//! # use std::{fs, io::Read};
 //! fn main() {
 //!     // read file into memory
-//!     let mut input = fs::File::open("sample.dtb").unwrap();
+//!     let mut input = fs::File::open("examples/bcm2709-rpi-2-b.dtb").unwrap();
 //!     let mut buf = Vec::new();
 //!     input.read_to_end(&mut buf).unwrap();
 //!
@@ -30,10 +31,14 @@
 //! }
 //! ```
 
-extern crate core;
+#![no_std]
+
+extern crate alloc;
+extern crate hashbrown;
 
 pub mod util;
 
+use alloc::{borrow::ToOwned, string::String, vec::Vec};
 use core::str;
 use util::{align, SliceRead, SliceReadError, VecWrite, VecWriteError};
 
@@ -131,7 +136,7 @@ impl From<str::Utf8Error> for DeviceTreeError {
 
 #[cfg(feature = "string-dedup")]
 mod advancedstringtable {
-    use std::collections::HashMap;
+    use hashbrown::HashMap;
 
     pub struct StringTable {
         pub buffer: Vec<u8>,
@@ -161,6 +166,8 @@ mod advancedstringtable {
 
 #[cfg(not(feature = "string-dedup"))]
 mod stringtable {
+    use alloc::vec::Vec;
+
     pub struct StringTable {
         pub buffer: Vec<u8>,
     }
@@ -512,35 +519,5 @@ impl From<str::Utf8Error> for PropError {
 impl From<SliceReadError> for PropError {
     fn from(e: SliceReadError) -> PropError {
         PropError::SliceReadError(e)
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use std::fs;
-    use std::io::{Read, Write};
-
-    use super::*;
-
-    #[test]
-    fn roundtrip() {
-        // read file into memory
-        let buf = include_bytes!("../examples/bcm2709-rpi-2-b.dtb");
-        let original_fdt = DeviceTree::load(buf).unwrap();
-
-        let dtb = original_fdt.store().unwrap();
-        let mut output = fs::OpenOptions::new()
-            .write(true)
-            .create(true)
-            .open("output.dtb")
-            .unwrap();
-        output.write_all(&dtb).unwrap();
-
-        let mut input = fs::File::open("output.dtb").unwrap();
-        let mut buf = Vec::new();
-        input.read_to_end(&mut buf).unwrap();
-        let generated_fdt = DeviceTree::load(buf.as_slice()).unwrap();
-
-        assert!(original_fdt == generated_fdt);
     }
 }
